Move SAND images to a subdirectory.
The one thing that annoys me most about using the SAND template is
that it has images in the base directory, and then I can't do open *.pdf
to open the document.  By putting all images in a subdirectory, the only
pdf file in the build directory will be the build document.