Changed Table environment to be left justified.
The function prototypes are now aligned with the left margin of the
page, making them also aligned with other source code and generally
look much nicer.  Also added LaTeX code to prevent breaking lines
and to use a ragged bottom if the line is broken (to prevent those
ugly wide spaces in the middle of the page).